chore: fade bottom tabs
